pyproject.toml
[tool.poetry]
name = "noiseworks"
version = "0.1.0"
description = ""
authors = ["mySociety"]
license = "AGPL"

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.3"
django-environ = "^0.9.0"
psycopg2 = "^2.8.6"
django-libsass = "^0.9"
crispy-forms-gds = "^0.2.2"
django-filter = "^22.1"
Authlib = "^1.0.1"
requests = "^2.25.1"
django-sesame = "^3.1"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^6.3.0"}
notifications-python-client = "^6.3.0"
django-formtools = "^2.3"
requests-cache = "<0.8"
django-simple-history = "^3.0.0"
boto3 = "^1.20.10"
smart-open = {extras = ["s3"], version = "^6.0.0"}
isort = "^5.10.1"
flake8 = "^5.0.4"

[tool.poetry.dev-dependencies]
black = "^22.1"
pytest-django = "^4.3.0"
django-debug-toolbar = "^3.2.1"
coverage = {extras = ["toml"], version = "^6.3"}
requests-mock = "^1.9.3"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "noiseworks.settings"
python_files = ["tests.py", "test_*.py"]
filterwarnings = [
    "ignore:'[^']*' defines default_app_config = '[^']*'. Django now detects this configuration automatically. You can remove default_app_config.:PendingDeprecationWarning",
    "ignore::django.utils.deprecation.RemovedInDjango40Warning",
]

[tool.coverage.run]
omit = ['*/asgi.py', '*/wsgi.py', 'manage.py']

[tool.black]
exclude = '''/(
  migrations
)/'''

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
